requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
beautifulsoup4
pydantic
Pillow
duckdb
pandas
markdown
pydub
pyaudio
flask
torch
pathlib
gitpython
numpy
speechrecognition
pyaudio
faker
